% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{BIC_GSS}
\alias{BIC_GSS}
\title{Compute BIC for a Gaussian graphical spike-and-slab model.}
\usage{
BIC_GSS(estimates, N)
}
\arguments{
\item{estimates}{A list of parameter estimates using the \code{navigm} function,
which includes the precision matrix (\code{Omega}), edge posterior inclusion probability (\code{m_delta}),
and a matrix \code{S (=Y^T Y)} evaluated based on data Y.}

\item{N}{number of samples, i.e., number of rows in the input data matrix Y of \code{navigm}.}
}
\value{
Scalar: Bayesian information criterion for a Gaussian graphical spike-and-slab model on data with N samples.
}
\description{
This function evaluates one of the model selection criteria, Bayesian information criterion (BIC),
for a Gaussian graphical spike-and-slab model.
}
\examples{
# simulate data

seed <- 123; set.seed(seed)
N <- 100; P <- 5; Q <- 3
V <- matrix(rnorm(P*Q), nrow = P, ncol = Q)
Y <- matrix(rnorm(N*P), nrow = N, ncol = P)

# estimate precision matrix based on Y and meanwhile leverage node-level variables V
ans <- navigm_core(Y, V)
BIC_GSS(ans$estimates,N)

}
