% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{compute_pauc}
\alias{compute_pauc}
\title{Compute standardised area under the truncated ROC curve.}
\usage{
compute_pauc(ppi, pat, fpr_stop = 1, standardise = F)
}
\arguments{
\item{ppi}{A vector of continuous prediction scores such as posterior inclusion probability or a list of such vectors.}

\item{pat}{A vector of binary true outcomes or a list of such vectors.}

\item{fpr_stop}{Scalar. False positive rate at which the ROC curve is truncated.}

\item{standardise}{Logical. If FALSE (default), not apply the standardisation; otherwise, apply.}
}
\value{
A scalar measuring the classification performance.
}
\description{
This function compute the area under the ROC curve (AUC) truncated at
the false positive rate \code{fpr_stop}. The AUC is a threshold-free performance measure.
}
