% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{compute_perf}
\alias{compute_perf}
\title{Compute thresholds-based performance measures.}
\usage{
compute_perf(ppi, pat, threshold = 0.5)
}
\arguments{
\item{ppi}{A vector of continuous prediction scores such as posterior inclusion probability or a list of such vectors.}

\item{pat}{A vector of binary true outcomes or a list of such vectors.}

\item{threshold}{Scalar. A threshold for classification.}
}
\value{
A list of thresholds-based performance measures.
}
\description{
This function compute the thresholds-based performance measures
including true positive rate (tpr), false positive rate (fpr), true negative rate (tnr),
false negative rate (fnr), recall, precision, F1-scores (f1).
}
