% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{EBIC_GSS}
\alias{EBIC_GSS}
\title{Evaluate the extended Bayesian information criterion (BIC) for spike-and-slab Gaussian graphical models.}
\usage{
EBIC_GSS(estimates, gamma = 0.5, N)
}
\arguments{
\item{estimates}{A list of parameter estimates using the \code{navigm} function,
which includes precision matrix (Omega), edge posterior inclusion probability (m_delta)
and a matrix S (=Y^T Y) evaluated based on data Y.}

\item{gamma}{EBIC parameter in [0,1] (default = 0.5). \code{gamma = 0} recovers BIC.
Positive gamma leads to stronger penalisation of large graphs.
https://arxiv.org/pdf/1011.6640.pdf shows \code{gamma=0.5} achieves a good compromise between positive selection rates and false discovery rates.}

\item{N}{number of samples (number of rows in Y).}
}
\value{
Scalar. Extended Bayesian information criterion of a spike-and-slab Gaussian graphical model on N-sample data Y.
}
\description{
This function evaluates one of the model selection criteria, extended Bayesian information criterion,
for spike-and-slab Gaussian graphical models.
}
