% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{generate_data_from_adjancency}
\alias{generate_data_from_adjancency}
\title{Simulate precision matrices and data given an adjacency matrix.}
\usage{
generate_data_from_adjancency(
  N,
  A,
  vec_magnitude = c(0.25, 0.75),
  bool_scale = TRUE
)
}
\arguments{
\item{N}{Scalar. Number of observations.}

\item{A}{An adjacency matrix.}

\item{vec_magnitude}{A vector of two positive numbers indicating the range of absolute magnitudes of precision matrix entries.}

\item{bool_scale}{Logical. If TRUE (default), scale the samples; otherwise, not scale.}
}
\value{
A list containing the simulated data:
 \describe{
 \item{A} the input adjacency matrix.
 \item{Omega} the simulated adjacency matrix of same size and structure of the pre-specified adjacency matrix.
 \item{Y} the simulated observations with N rows and \code{ncol(A)} columns.
}
}
\description{
This function simulates precision matrices and the observations based on
the pre-specified adjacency matrix,
inspired by Learning Graphical Models With Hubs, JMLR, 2014, P.3307.
}
