% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{generate_V}
\alias{generate_V}
\title{Simulate beta distributed auxiliary variables.}
\usage{
generate_V(
  P,
  Q,
  alpha = 0.05,
  beta = 0.2,
  Sigma = diag(Q),
  min_gene = round(0.05 * P),
  verbose = F
)
}
\arguments{
\item{P}{Scalar. Number of nodes in the graph.}

\item{Q}{Scalar. Number of node-level auxiliary variables.}

\item{alpha, beta}{Scalars. Shape parameters of beta distribution.}

\item{Sigma}{A correlation matrix of auxiliary variables.}

\item{min_gene}{Scalar. Number of influential entries per auxiliary variable.}

\item{verbose}{Logical. If FALSE (default), not show messages; otherwise, show the messages.}
}
\description{
This function simulates auxiliary variable matrices in order to assess the package.
}
