% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{get_fdr_threshold}
\alias{get_fdr_threshold}
\title{Compute a threshold based on a Bayesian FDR.}
\usage{
get_fdr_threshold(
  ppi,
  threshold_v = seq(1e-04, 0.99, length.out = 10^4 + 1),
  fdr = 0.05
)
}
\arguments{
\item{ppi}{A vector of continuous prediction scores, such as posterior inclusion probability, or a list containing such vectors.}

\item{threshold_v}{A vector of candidate thresholds. The default is \code{seq(1e-4, 0.99, length.out = 10^4 + 1)}.}

\item{fdr}{Scalar: the target Bayesian false discovery rate.}
}
\description{
This function calculates a threshold based on the Bayesian false discovery rate (FDR).
The method follows the approach outlined in Newton, M.A., Noueiry, A., Sarkar, D. and Ahlquist, P., 2004. Detecting differential gene expression with a semiparametric hierarchical mixture method. Biostatistics, 5(2), pp.155-176.
}
\examples{
seed <- 123; set.seed(seed); x <- pnorm(rnorm(100))
get_fdr_threshold(x)

}
