% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigss_vbem_core.R
\name{navigss_vbem_core}
\alias{navigss_vbem_core}
\title{Node-level auxiliary variable informed Gaussian spike-and-slab using variational EM.}
\usage{
navigss_vbem_core(
  Y,
  V = NULL,
  ne0 = NULL,
  method = "GMSS",
  list_hyper = NULL,
  list_init = NULL,
  tol = 0.1,
  maxit = 1000,
  seed = NULL,
  verbose = T,
  track_ELBO = F,
  debug = F,
  version = NULL,
  transformV = T
)
}
\arguments{
\item{Y}{Data matrix of dimension N x P, where N is the number of samples and P is the number of nodes in the graph.
\code{Y} will be centred within \code{navigss_vbem_core} call.}

\item{V}{Input matrix of dimension P x Q, where Q is the number of candidate auxiliary variables. No need to supply intercept.
\code{V} will be centred within \code{navigss_vbem_core} call if not within [0,1] by default, unless disabled by users via \code{transformV = F}.
If \code{V} is not specified, will use \code{method = 'GM'}.}

\item{list_hyper}{A list containing the model hyperparameters:
  the gamma prior on the continuous spike-and-slab scale requires shape parameter \code{a_tau} and rate parameter \code{b_tau};
  the unscaled spike and slab variances in the continuous spike-and-slab are v0 and v1;
  the exponential prior on diagonal entries requires rate parameter lambda;
  the normal prior on the overall sparsity requires mean n0 and variance t02;
  the gamma prior on the slab precision in the discrete spike-and-slab requires shape parameter \code{a_sigma} and rate parameter \code{b_sigma} (needed in GMN & GMSS);
  the beta prior on auxiliary variable inclusion probability requires shape parameters \code{a_o} and \code{b_o} (only needed in GMSS);
  the beta prior on edge inclusion probability requires shape parameters \code{a_rho} and \code{b_rho} (only needed in \code{method = "GM"} and \code{version = 1});

  if NULL or any hyperparameter specification is missing,
  set to default, \code{lambda = 2, v0 = 0.1, v1 = 100, a_tau = b_tau = a_sigma = b_sigma = 2, n0 = -2, t02 = 0.5, a_o = 1, b_o = Q, a_rho = 1, b_rho = P}.}

\item{list_init}{A list containing the initial variational parameters:
the gamma variational distribution of the continuous spike-and-slab scale includes shape parameter \code{alpha_tau} and rate parameter \code{beta_tau};
the normal variational distribution of the overall sparsity includes mean \code{mu_zeta} and variance \code{sig2_inv_zeta};
the normal variational distribution of regression coefficients includes mean \code{mu_beta} and variance \code{sig2_inv_beta} (needed in GMN & GMSS);
the gamma variational distribution of slab precision in the discrete spike-and-slab includes \code{a_sigma, b_sigma} (needed in GMN & GMSS);
the beta variational distribution of auxiliary variable inclusion probability includes \code{a_o} and \code{b_o} (only needed in GMSS);
the beta variational distribution of edge inclusion probability includes \code{a_rho} and \code{b_rho} (only needed in \code{method = "GM"} and \code{version = 1});
if \code{NULL} or any initialisation is missing,
set to default, \code{mu_beta}: vector of zeros, \code{sig2_inv_beta}: vector of ones, \code{a_tau = b_tau = a_sigma = b_sigma = 1}, \code{mu_zeta = list_hyper$n0},
\code{sig2_inv_zeta = 1/list_hyper$t02}, \code{a_o = 1, b_o = Q},  \code{a_rho = 1, b_rho = P}.}

\item{tol}{Scalar: tolerance for the stopping criterion (default is 0.1).}

\item{maxit}{Scalar: maximum number of iterations allowed (default is 1000).}

\item{seed}{Scalar: seed for reproducible default choices of initialisation (default is 123).}

\item{verbose}{Logical; if \code{TRUE} (default), standard verbosity; otherwise, no messages.}

\item{track_ELBO}{Logical; if \code{TRUE} (default), track ELBO after each maximisation step; otherwise, not track.}

\item{debug}{Logical; if \code{FALSE} (default), not record additional terms for the purpose of debug;
otherwise, track ELBO within each variational update and number of warnings of decreasing ELBOs, possibly due to numerical round-off.}

\item{version}{Integer; take values of 1 (a beta prior on edge inclusion) or 2 (a normal prior on probit edge inclusion) only. Only valid when \code{method = 'GM'}.}
}
\value{
A list containing the following estimates and settings:
 \describe{
 \item{estimates}{A list containing estimates of variational distribution parameters, which includes a selection of following terms depending on method:
 \code{Omega} (estimated precision matrix of dimension P x P),
 \code{m_delta} (posterior inclusion probability matrix of dimension P x P. off-diagonal entry (i, j) corresponds to
 the variational posterior probability of edge (i,j) being included in the graph),
 \code{m_gamma} (posterior inclusion probability vector of dimension Q. entry Q corresponds to the variational posterior probability of
          qth auxiliary variable explaining the node centrality),
 \code{mu_beta, sig2_inv_beta} (vector of length Q containing the variational posterior mean and inverse variance of regression coefficients of
          node-level auxiliary variables in columns of matrix \code{V}),
\code{a_tau, b_tau} (estimated parameters of Gamma distributed scale in continuous spike-and-slab),
\code{a_sigma, b_sigma} (estimated parameters of Gamma distributed slab variance in discrete spike-and-slab,
\code{a_o, b_o} (estimated parameters of Beta distributed auxiliary variable inclusion probability),
\code{mu_zeta, sig2_inv_zeta} (variational posterior mean and inverse variance of overall sparsity parameter).}
 \item{debugs}{A list containing terms for debugging:
 \code{n_warning} (scalar. number of ELBO drops in the maximisation step),
 \code{vec_n_warning_VB} (vector of length it. number of ELBO drops in each variational step),
 \code{list_elbo} (a list containing the ELBO in each variational step).  }
 \item{args}{A list containing the algorithm inputs including \code{list_hyper} and \code{list_init}.}
\item{it}{Scalar. Final number of iterations. \code{it == maxit} indicates the algorithm dose not converge. }
\item{vec_ELBO_M}{Vector of length \code{it}. ELBOs after the maximisation step. }
\item{vec_VB_it}{Vector of length \code{it}. Number of iterations in each variational step. \code{any(vec_VB_it == maxit)} indicates the variational step dose not converge.}
\item{pt}{Scalar. Algorithm runtime in seconds. }
}
}
\description{
This function performs variational Bayes expectation maximization for
simultaneous Gaussian spike-and-slab inference and node-level auxiliary variable selections.
}
\details{
\code{navigss_vbem_core} implements a Gaussian graphical model
  that allows incorporating and selecting node-level auxiliary variables,
  thereby enhancing the detection of conditional dependence. Inference is
  carried out using a scalable variational expectation maximisation
  algorithm, which is applicable to high-dimension graphs.
}
\examples{
# simulate data

seed <- 123; set.seed(seed)
N <- 100; P <- 5; Q <- 3
V <- matrix(rnorm(P*Q), nrow = P, ncol = Q)
Y <- matrix(rnorm(N*P), nrow = N, ncol = P)

# estimate precision matrix based on Y and meanwhile leverage node-level variables V
res_navigss_vbem_core <-navigss_vbem_core(Y, V)
# res_navigss_vbem_core <-navigss_vbem_core(Y, V, method = 'GMN')
# res_navigss_vbem_core <-navigss_vbem_core(Y, V, method = 'GM', version = 1)
# res_navigss_vbem_core <-navigss_vbem_core(Y, V, method = 'GM', version = 2)

}
