% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{plot_ppi}
\alias{plot_ppi}
\title{Plot variable-specific quantities such as effect sizes and PPIs.}
\usage{
plot_ppi(
  ppi,
  ppi_names = NULL,
  col = "black",
  condition = (ppi >= 0.5),
  xlab = "auxiliary variables",
  ylab = "effect sizes",
  ...
)
}
\arguments{
\item{ppi}{A vector of the variable-specific quantities.}

\item{ppi_names}{A vector of variable names. If \code{NULL} (default), use the variable indices.}

\item{col}{A character indicating the point color or a vector of the same length of variables.}

\item{condition}{A vector of logical. If satisfying \code{condition == T}, draw vertical lines from points to the x-axis.}

\item{xlab}{Character. A title for the x axis.}

\item{ylab}{Character. A title for the y axis.}

\item{...}{other plotting arguments. See \code{base::plot}.}
}
\description{
This function plots variable-specific quantity \code{ppi} against their indices \code{ppi_names},
for instance posterior inclusion probabilities or effect sizes of auxiliary variables
}
