% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot truncated receiver operating characteristic curves.}
\usage{
plot_roc(ppi, pat, fpr_stop = 1, nci = 11, ...)
}
\arguments{
\item{ppi}{A vector of continuous prediction scores such as posterior inclusion probability or a list of such vectors.}

\item{pat}{A vector of binary true outcomes or a list of such vectors.}

\item{fpr_stop}{Scalar. False positive rate at which the ROC curve is truncated.}

\item{nci}{number of confidence bars if \code{ppi} and \code{pat} are lists.}

\item{...}{other plotting arguments, such as \code{col}, \code{add}. Full details see \code{ROCR::plot}.}
}
\description{
This function plots the truncated receiver operating characteristic (ROC) curves at
the false positive rate \code{fpr_stop}. ROC curves describe the
performance of a classification model at all classification thresholds.
}
