% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{simulate_auxiliary_matrix}
\alias{simulate_auxiliary_matrix}
\title{Simulate an auxiliary variable matrix.}
\usage{
simulate_auxiliary_matrix(
  P,
  Q,
  alpha = 0.05,
  beta = 0.2,
  Sigma = diag(Q),
  min_gene = round(0.05 * P),
  verbose = F
)
}
\arguments{
\item{P}{Scalar: number of nodes in the graph.}

\item{Q}{Scalar: number of node-level auxiliary variables.}

\item{alpha, beta}{Scalars: shape parameters of the beta distribution (default values: \code{alpha = 0.05} and \code{beta = 0.2}).}

\item{Sigma}{A matrix encoding the correlation between auxiliary variables (default value: an identity matrix).}

\item{min_gene}{Scalar: minimum number of nodes influenced by each auxiliary variable.}

\item{verbose}{Logical: If set to FALSE (default), messages will not be displayed; if set to TRUE, messages will be shown.}
}
\value{
A matrix of simulated auxiliary variables with dimensions \code{P x Q}.
}
\description{
This function simulates an auxiliary variable matrix.
Auxiliary variables mimic the posterior probability of inclusion in a Bayesian sparse regression with spike-and-slab priors on regression coefficients.
The entries are generated from a right-skewed beta distribution.
}
\examples{
simulate_auxiliary_matrix(P = 50, Q = 10)

}
