% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{simulate_data_from_adjacency_matrix}
\alias{simulate_data_from_adjacency_matrix}
\title{Simulate precision matrices and data based on an adjacency matrix.}
\usage{
simulate_data_from_adjacency_matrix(
  N,
  A,
  vec_magnitude = c(0.25, 0.75),
  bool_scale = TRUE
)
}
\arguments{
\item{N}{Scalar: number of observations.}

\item{A}{An adjacency matrix.}

\item{vec_magnitude}{A vector containing two positive numbers
that indicate the range of absolute magnitudes for the off-diagonal entries in the precision matrix.}

\item{bool_scale}{Logical: If set to TRUE (default), the samples will be scaled; otherwise, they will not be scaled.}
}
\value{
A list containing the simulated data:
 \describe{
\item{A}{The input adjacency matrix.}
\item{Omega}{The simulated precision matrix of the same size and structure as the pre-specified adjacency matrix.}
\item{Y}{The simulated observations with \code{N} rows and \code{nrow(A) = ncol(A)} columns.}
}
}
\description{
This function simulates precision matrices that adhere to the structure of a given adjacency matrix and
generates observations from a multivariate normal distribution with mean 0 and the simulated precision matrix.
The procedure follows "Learning Graphical Models With Hubs," JMLR, 2014, p. 3307.
}
\examples{
A <- matrix(0, 5, 5); A[1,2] <- A[2,1] <- A[2,5] <- A[5,2] <- 1;
simulate_data_from_adjacency_matrix(N = 10, A = A)

}
