% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{BIC_GSS}
\alias{BIC_GSS}
\title{Evaluate the Bayesian information criterion (BIC) for spike-and-slab Gaussian graphical models.}
\usage{
BIC_GSS(estimates, N)
}
\arguments{
\item{estimates}{A list of parameter estimates using the \code{navigm} function,
which includes precision matrix (Omega), edge posterior inclusion probability (m_delta)
and a matrix S (=Y^T Y) evaluated based on data Y.}

\item{N}{number of samples (number of rows in Y).}
}
\value{
Scalar. Bayesian information criterion of a spike-and-slab Gaussian graphical model on N-sample data Y.
}
\description{
This function evaluates one of the model selection criteria, Bayesian information criterion,
for spike-and-slab Gaussian graphical models.
}
