% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{compute_pauc}
\alias{compute_pauc}
\title{Compute AUC.}
\usage{
compute_pauc(ppi, pat, fpr_stop = 1, standardise = F)
}
\arguments{
\item{ppi}{A vector of continuous prediction scores, such as posterior inclusion probability, or a list containing such vectors.}

\item{pat}{A vector of binary true outcomes, or a list containing such vectors.}

\item{fpr_stop}{Scalar: false positive rate at which the ROC curve is truncated.}

\item{standardise}{Logical: if set to FALSE (default), standardisation will not be applied;
if set to TRUE, standardisation will be applied following https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3744586/.}
}
\value{
Scalar: (standardised partial) area under the ROC curve.
}
\description{
This function computes the area under the ROC curve (AUC).
A partial AUC can be specified with the argument \code{fpr_stop}.
}
\examples{
set.seed(123); pat <- sample(c(0,1), 10, replace = TRUE)
ppi <- sapply(pat,function(x){if(x==0){runif(1,0,0.1)}else{runif(1,0.9,1)}})
compute_pauc(ppi, pat)

}
