% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{generate_data_from_adjancency}
\alias{generate_data_from_adjancency}
\title{Simulate precision matrices and data given an adjacency matrix.}
\usage{
generate_data_from_adjancency(
  N,
  A,
  vec_magnitude = c(0.25, 0.75),
  bool_scale = TRUE
)
}
\arguments{
\item{N}{Scalar. Number of observations.}

\item{A}{An adjacency matrix.}

\item{vec_magnitude}{A vector of two positive numbers indicating the range of absolute magnitudes of precision matrix entries.}

\item{bool_scale}{Logical. If TRUE (default), scale the samples; otherwise, not scale.}
}
\value{
Input adjacency matrix \code{A}, simulated precision matrix \code{Omega}, observations \code{Y}.
}
\description{
This function simulates precision matrices and the observations based on pre-specified adjacency matrix, inspired by Learning Graphical Models With Hubs, JMLR, 2014, P.3307.
}
