% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{plot_ppi}
\alias{plot_ppi}
\title{Plot auxiliary-variable-associated quantities.}
\usage{
plot_ppi(
  ppi,
  ppi_names = NULL,
  col = "black",
  condition = (ppi >= 0.5),
  xlab = "auxiliary variables",
  ylab = "effect sizes",
  ...
)
}
\arguments{
\item{ppi}{A vector of the values associated with auxiliary variables, such as effect sizes and posterior probability of inclusion.}

\item{ppi_names}{A vector of variable names. If set to \code{NULL} (default), the indices \code{1:length(ppi)} will be used.}

\item{col}{A character indicating the color of all the points, or a vector containing such characters with the same length as the variables.}

\item{condition}{A vector of logical values. If \code{condition == TRUE}, vertical lines will be drawn from points to the x-axis.}

\item{xlab}{Character: a title for the x-axis (the default is "auxiliary variables").}

\item{ylab}{Character: a title for the y axis (the default is "effect sizes").}

\item{...}{Other plotting arguments. Refer to \code{base::plot} for more details.}
}
\description{
This function generates a scatter plot illustrating the values associated with auxiliary variables \code{ppi} against their corresponding indices \code{ppi_names}.
}
\examples{
set.seed(123); pat <- sample(c(0,1), 10, replace = TRUE)
ppi <- sapply(pat,function(x){if(x==0){runif(1,0,0.1)}else{runif(1,0.9,1)}})
plot_ppi(ppi, ylab = 'inclusion probability')

}
