% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot pROC curve with 95\% confidence interval.}
\usage{
plot_roc(ppi, pat, fpr_stop = 1, nci = 11, spread.scale = 2, ...)
}
\arguments{
\item{ppi}{A vector of continuous prediction scores, such as posterior inclusion probability, or a list containing such vectors.}

\item{pat}{A vector of binary true outcomes, or a list containing such vectors.}

\item{fpr_stop}{Scalar: false positive rate at which the ROC curve is truncated.}

\item{nci}{Number of confidence bars. The 95\% confidence interval is only displayed if \code{ppi} and \code{pat} are lists.}

\item{...}{Other plotting arguments, such as \code{col} and \code{add}. For full details, please refer to \code{ROCR::plot}.}
}
\description{
This function plots an average receiver operating characteristic (ROC) curve with 95% confidence interval.
A partial ROC (pROC) curve can be specified with the argument \code{fpr_stop}.
}
\examples{
set.seed(123); pat <- sample(c(0,1), 10, replace = TRUE)
ppi <- sapply(pat,function(x){if(x==0){runif(1,0,0.1)}else{runif(1,0.9,1)}})
plot_roc(ppi, pat)

}
